Action rewrite
>500 lines to <150. 2 years progress makes a difference. This rewrite feels super basic, and that's cause it is. It's super easy (now) to do this compared to 2 years ago. I'm still writing 500 line things, but they're not boring-to-edit super repetitive things.

2 years goes from >500 lines for a stupid basic sentence generator to >500 lines for a feature-rich Game of Life web app with saving, loading, downloading, uploading, and more

Still maybe 2 years ago I shouldn't have wrote that bad? It was literally just a bunch of if/else statements and a bit of string manipulation, as in the tiniest bit. I remember thinking there was probably a way to do what this commit did 2 years ago, but I didn't know what that way would be. Ahh, the time before I knew what objects or arrays were. Not a very good time actually. Now is much better.